backend/requirements.txt
africastalking==2.0.2
arabic-reshaper==3.0.0
asgiref==3.11.1
asn1crypto==1.5.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
cssselect2==0.9.0
dj-database-url==3.1.2
Django==6.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.7.0
django-filter==25.2
freetype-py==2.5.1
html5lib==1.1
idna==3.11
lxml==6.0.2
openpyxl==3.1.5
oscrypto==1.3.0
pillow==12.1.1
psycopg2-binary==2.9.11
pycairo==1.29.0
pycparser==3.0
pyHanko==0.34.1
pyhanko-certvalidator==0.30.1
PyJWT==2.12.1
pypdf==6.9.1
python-bidi==0.6.7
python-dotenv==1.2.2
PyYAML==6.0.3
reportlab==4.4.10
requests==2.32.5
responses==0.26.0
rlPyCairo==0.4.0
schema==0.7.8
six==1.17.0
sqlparse==0.5.5
svglib==1.6.0
tinycss2==1.5.1
tzdata==2025.3
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
webencodings==0.5.1
xhtml2pdf==0.2.17
